# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central management server for the OpenLabStats fleet. Manages agent enrollment, software mappings, lab groupings, and reporting.",
        "title": "OpenLabStats Server API",
        "contact": {},
        "version": "0.1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/agents/rollout-status": {
            "get": {
                "description": "Per-platform rollout progress: how many agents are updated, currently updating, or pending, plus the current version distribution and target.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Agent auto-update rollout status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/installers/generate": {
            "post": {
                "description": "Returns a download URL for the latest pre-built MSI with an msiexec command.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installers"
                ],
                "summary": "Generate a custom MSI installer",
                "parameters": [
                    {
                        "description": "Installer configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GenerateInstallerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.GenerateInstallerRequest": {
            "type": "object",
            "properties": {
                "building": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                },
                "serverAddress": {
                    "type": "string"
                }
            }
        },
...
    }
}